update flag encoding and decoding to use '+' instead of '$' and update os separator used
And break the WinFormsUI project in the process. Anyway, when running the randomizer in Linux, it breaks if the flag contains a dollar sign, since Linux doesn't interpret it as part of a whole string. It can be fixed with passing the flags as single quotes, but after some discussions with Ellendar we finally decided to remove the dollar sign altogether and replace it with a plus '+' symbol. The os separator being used was also updated since previously the backslash was being used, but Linux uses slash, so now the new ROM can be saved in the same directory as the original ROM in both Windows and Linux (and in theory every other OS). With this commit, the CLI works from start to finish. My idea is to fix the code that broke in the WinFormsUI project in the next commit.