Change the way level up cancelling works if level cap is less than 8
If the level cap for a stat is less than 8, we do not want to allow
the player to cancel out of the level up screen to continue chasing
the experience for that stat once it is maxed. To stay true to vanilla,
this is still allowed if the level cap is 8.

The point of this is to be able to have lower max stat levels, but
not have this be abusable to reach high amounts of exp quickly using
crystals.