Stop losing all settings when RandomizerConfiguration has changed
JsonIgnore is added to RandomizerConfiguration.Flags so that we do not try to deserialize the previous flag string on startup. The previous flag string is not required here anyway because all the configurable fields are already serialized. The Flags field would likely fail deserialization when some field had changed, which invalidated all of the stored settings.