Add option to set Minimum Path To Dark Link (number of rooms)
This is proposed as a solution to the anticlimactic, very short GP's that sometimes happen.

How it works is it calculates the shortest path to Dark Link and invalidates the palace if it is shorter than the selected option. This adds some seed generation time, but not too much (plus it's optional.)