Nerf Minimum Path To Dark Link Enum
I want the short value to match the beginner preset, and the medium value to match standard. (I don't think it's desirable for beginners to have very short GP's either, so I like having it set for the beginner preset, but shorter.)

The values are geared towards coordinate palaces though. For Reconstructed the impact is more-or-less doubled, because of how they connect.