Fix so you can edit the drop pool without adding in random items
This was possible in 4.3, so I consider this a bug fix. You could only select randomize (which force-added other items until 8 drop items) or vanilla.

Additionally, the logic for small and large drop pools are separated now so you can keep either the small or large pool vanilla by selecting nothing in that group. This is useful if you want the vanilla small drop pool (1/8 chance for 50 P-bag) and only want to change the large drop pool, for example.