Fix bug with segmented item rooms
Sideviews for both the main and linked rooms were added to the bank. This would normally not cause any issue (as js65 would optimize them to one anyway). However, for item rooms the bytes are typically modified for the main room, but not the linked room. If the linked room was then written after the main room, it would then overwrite the entry for the main room.

Anyway there's no point to write the linked room sideview bytes, so I changed it to only output sideviews of enabled rooms.

Tested on the original reported seed and it works now.